Scoot the intro Panels into a subpackage of "houses" and move the HouseInformationPage into the "houses" package
