+++ src/main/java/website/fgto/houses/HouseInformationPage.java
@@ -1,9 +1,11 @@
-package website.fgto;
+package website.fgto.houses;
 
 import org.apache.wicket.markup.html.WebPage;
 import org.apache.wicket.markup.html.basic.Label;
 import org.apache.wicket.request.mapper.parameter.PageParameters;
-import website.fgto.houses.*;
+import website.fgto.Apology;
+import website.fgto.Footer;
+import website.fgto.houses.intros.*;
 
 public class HouseInformationPage extends WebPage {
   private static final long serialVersionUID = 1L;
