+++ src/main/java/website/fgto/houses/intros/Bruce.java
@@ -1,11 +1,11 @@
-package website.fgto.houses;
+package website.fgto.houses.intros;
 
 import org.apache.wicket.markup.html.panel.Panel;
 
-public class Wessex extends Panel {
+public class Bruce extends Panel {
   private static final long serialVersionUID = 1L;
 
-  public Wessex(String id) {
+  public Bruce(String id) {
     super(id);
   }
 }
