+++ src/main/java/website/fgto/houses/intros/CenelnGabran.java
@@ -1,11 +1,11 @@
-package website.fgto.houses;
+package website.fgto.houses.intros;
 
 import org.apache.wicket.markup.html.panel.Panel;
 
-public class Stuart extends Panel {
+public class CenelnGabran extends Panel {
   private static final long serialVersionUID = 1L;
 
-  public Stuart(String id) {
+  public CenelnGabran(String id) {
     super(id);
   }
 }
