+++ src/main/java/website/fgto/houses/intros/Normandy.java
@@ -1,11 +1,11 @@
-package website.fgto.houses;
+package website.fgto.houses.intros;
 
 import org.apache.wicket.markup.html.panel.Panel;
 
-public class Plantagenet extends Panel {
+public class Normandy extends Panel {
   private static final long serialVersionUID = 1L;
 
-  public Plantagenet(String id) {
+  public Normandy(String id) {
     super(id);
   }
 }
