+++ src/main/java/website/fgto/houses/intros/Stuart.java
@@ -1,11 +1,11 @@
-package website.fgto.houses;
+package website.fgto.houses.intros;
 
 import org.apache.wicket.markup.html.panel.Panel;
 
-public class Tudor extends Panel {
+public class Stuart extends Panel {
   private static final long serialVersionUID = 1L;
 
-  public Tudor(String id) {
+  public Stuart(String id) {
     super(id);
   }
 }
