+++ src/main/java/website/fgto/houses/intros/Wessex.java
@@ -1,11 +1,11 @@
-package website.fgto.houses;
+package website.fgto.houses.intros;
 
 import org.apache.wicket.markup.html.panel.Panel;
 
-public class CenelnGabran extends Panel {
+public class Wessex extends Panel {
   private static final long serialVersionUID = 1L;
 
-  public CenelnGabran(String id) {
+  public Wessex(String id) {
     super(id);
   }
 }
