+++ src/main/java/website/fgto/houses/intros/Windsor.java
@@ -1,11 +1,11 @@
-package website.fgto.houses;
+package website.fgto.houses.intros;
 
 import org.apache.wicket.markup.html.panel.Panel;
 
-public class Hanover extends Panel {
+public class Windsor extends Panel {
   private static final long serialVersionUID = 1L;
 
-  public Hanover(String id) {
+  public Windsor(String id) {
     super(id);
   }
 }
